Use cli flags, wrap main for defers.
Now we actually disconnect from dcrd when CTRL+C is hit, and wait for any calls to updatetemplateInformation to complete.

Since http.ListenAndServe blocks and we can't have it listen for a signal, we now fire it off and let it go.  But if it fails to bind, it will signal the other handlers to quit.

Also address linters.

Uncomment updatetemplateInformation.

fix grammar