New agendadb package for storing agendas and choices.
AgendaTagged is identical to dcrjson.Agenda except for tagging.  The primary key is "Id", and several fields are tagged as indexes for fast lookup.

ChoiceLabeled embeds dcrjson.Choice along with the AgendaID for the choice, and a string array suitable for use as a primary key. The AgendaID is tagged as an index for quick lookups based on the agenda.

Make tests to verify they will work for us, and as examples for the app.  A test agenda is saved and loaded by primary key.  Two test choices are saved and loaded by primary key.  A third is saved with a different AgendaID, and and the two with the common AgendaID are loaded.  Updating the value of a single field is also tested.