Show vote results for historic agendas. (#209)
The agenda details on the html template are populated using an Agenda
object. This PR removes the logic which creates that Agenda object, and
replaces it with logic which can create an Agenda object for **every**
agenda, not just the current one.

The code favours readability and maintainability over runtime efficiency, as it will only be executed once at startup and once each time a block is discovered.